Fixed Death Glitch
23/12/14

Removed north side dog monster and replaced with with the middle side dog. Killing that north side dog monster spawned the player at the origin point of trenchbroom and froze the player in shooting animation. They cannot move but can still jump and turn the camera around. Something must have happened when creating the trees that made the dog glitch or get corrupted and doing that to the player. Still unknown as to what reasons could have cause this since nothing about the dog was modified.